pwm: Correct the "request" SYSFS interface
This patch changes handling of "request" SYSFS interface
1. Request for PWM device
from "cat request" to "echo 1 > request"
2. Release for PWM device
from "echo 1 > request" to "echo 0 > request"
3.Information on reading "request" SYSFS interface
if pwm device used, "<pwm device> is used by which <application>"
or if pwm device is free, "<pwm device> is free"

Also this patch has the fix for hanging of the system on subsequent
request/release operation after release of unused device. This is
because, release of unused device causes mutex to hold in a locked state
and further request for the mutex would hang.

Signed-off-by: Philip, Avinash <avinashphilip@ti.com>